Fix deleting a dictionary
When the user presses the "delete" button in the Add-on dictionary
interface, we send an empty dictionary to Latin IME as a dummy for
it to use instead of the pre-installed one.
The old "empty" placeholder was rejected by initial verification.
This one is still empty, but passes the check.

Bug: 7980985
Change-Id: I751bbf67d8078dd5fbdce1dd2f0db72a8207816d