Modify isAllowedToLog
Currently isAllowedToLog only checks the state of the
ResearchLogger, and does not consider which log the data is
going to.  This causes problems with the logs for Feedback.
The Feedback log should inhibit normal logging procedures,
but the system needs to be able to write specific data
directly to it.

This change renames to isAllowedToLogTo and adds the
destination ResearchLog as a parameter.  A FeedbackLog is also
added as a new class so it can be distinguished from other
ResearchLogs.

Change-Id: I5a1eea05bb040c26bf816b89179f44b3024fa2ad