Fix ResearchLogger feedback bug
If the logger feedback screen is up and the user hits the home button or the
keyboard otherwise closes, the system won't let the user enter feedback again,
because it is waiting for the user to finish the first one.

With this change, the state associated with the feedback screen is
cleared when the user closes the keyboard.

b/9398772

Change-Id: I832a82793ad42e31b6b3fdb04fabcea75facde7d