Fix the suggestion size.
The available space was not computed correctly. This makes the
decision to ellipsize/expand the middle suggestion correct.

Also, the scale was not applied correctly in some cases.

Finally, for some reason ellipsize does not work as we expect.
I'm not sure how to get it to work like we want it to, but
this works around the problem in a simple manner.

Bug: 11158757
Change-Id: I083e7ac56e157f100923733e67b92a6368a526c9