Fix messages in tests
They are not failing because of this right now, but future
tests would fail without this.
Concretely, since the Handler for any given main thread is
a singleton managed by the framework, the messages won't
be removed automatically. In particular, messages posted
with a delay have a large destructive potential, because
they will go through before the startInputViewInternal
method is called, because it's called with a message too (in
turn, this is because of working around framework bugs).

Bug: 11142685
Change-Id: If81a168efcfd7d93806ddc4ea432d438c45a3e1d