Disable empty suggestion word on the suggestions strip
It should be disabled when it is empty. But should be enabled even
when it is empty with accessibility exploration on to avoid announcing
"disabled".

Note that this CL may not be able to handle "suspending talk back"
state correctly. It should be handled by registering
TouchExplorationStateChangeListener that is available from API 19.

Bug: 15738513
Bug: 14849593
Change-Id: Id9ad89f770840f74a106f53833060407b20c1e24