Use FrameLayout for the backing view
If we render our UI around the composing text in the fullscreen
mode, probably we need to have our views under
android.R.id.extractArea, which is implemented as FrameLayout.
Therefore, using FrameLayout for the backing view makes it easy
to simplify the code between normal mode and full screen mode.

Change-Id: I69533b22d9d1368ad906dc5ef7be1a1b674d10b8