Add !string/<resource_name> reference
This CL introduces new text reference notation !string/<resource_name>
to refer a string resource on the fly.

This notation is mainly used to represent action key labels may refer
a string in a system locale in run-time.

This notation is needed to implement Hinglish and Serbian-Latin
keyboards that need to refer its own action key labels.

Bug: 17169632
Bug: 9687668
Change-Id: I042f6bd04714e0e448cd92031730eb9fb422e6d3