[ML5] Rewrite resetDictionariesWithDictNamePrefix
This change rewrites resetDictionariesWithDictNamePrefix so
that it would work if there were several locales, which is going to
happen shortly. This change on its own is quite large and difficult
already, so I didn't want to merge it into the change that actually
makes several locales happen; another such change is following.

Change-Id: Ibb80b05b73c0f0f0bea64b7e308cc4d38ca448f9