Tune keyboard height and letter size of tablet
The height of Google Keyboard is limited by maximum and minimum ratio
to the display height.  Because screen aspect ratio of some large
tablet is close to 1:1, the height of the keyboard may become taller
than expected. This CL lowers the minimum ratio to allow shorter
height. This CL also increase text size of Material theme keyboard on
tablet as well.

Bug: 17050390
Change-Id: I87c5f49ce554950c525a1f31dfddfd1beb3fafda