Support CursorAnchorInfo in the full-screen mode
This CL allows to emulate the behavior of CursorAnchorInfo
in the full-screen mode to work around the limitation of
the framework.

Basically this CL copies relevant code from the framework
to LatinIME. This is actually a code duplicate which isn't
great, but probably the only way that is available right now.

Change-Id: I30bc8c3387c91c8e47252ba699cbd967d8927dad