Hide app icon when keyboard is a system app.
The objective is to show an app icon when the user installs the keyboard from
the Play Store, but not show the icon when it comes from the system image.
The latter rule applies to AOSP keyboards and to the Google Keyboard on a
Nexus device.

This change partially reverts ag/613096.

Bug 19001197.

Change-Id: Icb202deffe9f1ab2bde9b3c9d221347da6793a19